backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0
streamlit==1.32.0
openai==1.3.7
requests==2.31.0
qrcode[pil]==8.2 